Refactoring, Initial version of the database logic.
