+++ source/Roy.Logging/Roy.Logging/Domain/Settings/Database/DatabaseSetting.cs
@@ -1,0 +1,8 @@
+﻿namespace Roy.Logging.Domain.Settings.Database;
+
+/// <summary>
+/// Database settings.
+/// </summary>
+public class DatabaseSetting
+{
+}
