+++ source/Roy.Logging/Roy.Logging/ExceptionExtension.cs
@@ -4,6 +4,7 @@ using Roy.Logging.Domain.Attributes;
 using Roy.Logging.Domain.Contants;
 using Roy.Logging.Domain.Program;
 using Roy.Logging.Domain.Settings;
+using Roy.Logging.Services;
 using System.Diagnostics;
 
 namespace Roy.Logging;
@@ -277,7 +278,7 @@ public static class ExceptionExtension
                     setting.Exception.LoadInformationSettings, 
                     webApplicationHttpContext, listOfParameters);
 
-                new RecordService().SaveAsync(exceptionDetail, 
+                new Record().SaveAsync(exceptionDetail, 
                     setting.Exception);
             }
         }
